# Digest ZSTD compression dictionary once when writing SST file