--- a/table/block_fetcher.cc
+++ b/table/block_fetcher.cc
@@ -257,9 +257,7 @@ Status BlockFetcher::ReadBlockContents() {
   if (do_uncompress_ && compression_type_ != kNoCompression) {
     // compressed page, uncompress, update cache
     UncompressionContext context(compression_type_);
-    CompressionDict dict;
-    dict.Init(compression_dict_, CompressionDict::Mode::kUncompression,
-              compression_type_);
+    UncompressionDict dict(compression_dict_, compression_type_);
     UncompressionInfo info(context, dict, compression_type_);
     status_ = UncompressBlockContents(info, slice_.data(), block_size_,
                                       contents_, footer_.version(), ioptions_,