… tools/db_bench_tool.cc
… utilities/blob_db/blob_db_impl.cc
… utilities/blob_db/blob_dump_tool.cc