This is essentially a re-submission of #4251 with a few improvements:

- Split `CompressionDict` into two separate classes: `CompressionDict` and `UncompressionDict`
- Eliminated `Init` functions. Instead do all initialization work in constructors.
- Added test case for parallel DB open, which is the scenario where #4251 failed under TSAN.

Test Plan:

- ZSTD version testing covering v0.4.7, v0.5.0, v0.6.2, v0.7.0, v1.1.2, v1.1.3, v1.3.8. Ran below `db_stress` command on each version, with dictionary trainer enabled for v1.1.3+.

```
$ TEST_TMPDIR=/data/compaction_bench LD_LIBRARY_PATH="/data/users/andrewkr/zstd/lib:$LD_LIBRARY_PATH" python tools/db_crashtest.py blackbox --max_key=10000000 --value_size_mult=33 --compression_type=zstd --write_buffer_size=1048576 --target_file_size_base=1048576 --max_bytes_for_level_base=4194304 --max_background_compactions=4 --interval=15 --duration=60 --compression_max_dict_bytes=16384 --compression_zstd_max_train_bytes=0 --checkpoint_one_in=0
```

- measured random write performance before/after this PR with ZSTD 1.3.8

benchmark command:
```
$ TEST_TMPDIR=/data/compaction_bench LD_LIBRARY_PATH="/data/users/andrewkr/zstd/lib:$LD_LIBRARY_PATH" ./db_bench -num=10000000 -max_write_buffer_number=4 -benchmarks=filluniquerandom --compression_type=zstd --write_buffer_size=4194304 --target_file_size_base=4194304 --max_bytes_for_level_base=16777216 --block_size=16384 --max_background_jobs=12 --compression_max_dict_bytes=$max_dict_bytes --compression_zstd_max_train_bytes=$zstd_max_train_bytes
```

results before this PR:
```
max_dict_bytes | zstd_max_train_bytes | tput
0 | 0 | 37.5 MB/s
16384 | 0 | 36.4 MB/s
16384 | 131072 | 36.4 MB/s
```

results after this PR:
```
max_dict_bytes | zstd_max_train_bytes | tput
0 | 0 | 39.8 MB/s
16384 | 0 | 38.9 MB/s
16384 | 131072 | 39.7 MB/s
```

- measured random read performance before/after this PR with ZSTD 1.3.8

setup command:
```
$ TEST_TMPDIR=/dev/shm LD_LIBRARY_PATH="/data/users/andrewkr/zstd/lib:$LD_LIBRARY_PATH" ./db_bench -num=10000000 -max_write_buffer_number=4 -benchmarks=filluniquerandom,readrandom --compression_type=zstd -
-write_buffer_size=4194304 --target_file_size_base=4194304 --max_bytes_for_level_base=16777216 --max_background_jobs=12 --compression_max_dict_bytes=$max_dict_bytes --compression_zstd_max_train_bytes=$zstd_max_train_bytes --bloom_bits=8
```

benchmark command:
```
$ TEST_TMPDIR=/dev/shm LD_LIBRARY_PATH="/data/users/andrewkr/zstd/lib:$LD_LIBRARY_PATH" ./db_bench -use_existing_db=true -num=10000000 -duration=600 -threads=32 -max_write_buffer_number=4 -benchmarks=readr
andom --compression_type=zstd --write_buffer_size=4194304 --target_file_size_base=4194304 --max_bytes_for_level_base=16777216 --max_background_jobs=12 --compression_max_dict_bytes=$max_dict_bytes --compression_zstd_max_train_bytes=$zstd_max_train_bytes --bloom_bits=8 -cache_size=0
```

results before this PR:
```
max_dict_bytes | zstd_max_train_bytes | tput (MB/s)
0 | 0 | 223.5
16384 | 0 | 220.7
16384 | 131072 | 183.3
```

results after this PR:
```
max_dict_bytes | zstd_max_train_bytes | tput (MB/s)
0 | 0 | 224.9
16384 | 0 | 200.2
16384 | 131072 | 173.6
```