@siying 2019-01-07
Do I understand correctly: after the change, digested dictionary is always used when supported? It means that the current users of compression dictionary will see memory usage increased dramatically?