@ajkr 2019-01-07
Yes, it is always used. With this PR we create digested dictionary in the following cases:

- when `BlockBasedTableBuilder` is constructed. These digested dictionaries live until the end of the builder's lifetime.
- before calling `UncompressBlockContents`. These digested dictionaries live only until a particular block is uncompressed.

I believe what you're worried about is if we held a digested dictionary for every `BlockBasedTableReader`. This PR doesn't do that yet because, like you said, it'd increase memory dramatically. I am looking into a way to put these digested dictionaries in block cache, though that'll be in the next PR.