@ajkr 2019-01-09
The perf testing for randomread doesn't look good. I'll investigate it before proceeding with this PR.