@ajkr 2019-01-09
Updated benchmark commands to remove potential sources of bias, and ran them much longer (10 minutes per run). Now randomread perf does not look regressed at all in the common case, i.e., no dictionary. I am ok with the regression in the dictionary case since it'll be fixed in a subsequent PR.